requirements.txt
pandas
scikit-learn
matplotlib
shap
numpy
umap-learn